+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Base properties that are hid and have different accessors are new explicitly implemented (issue [#227](https://github.com/JasonBock/Rocks/issues/227))
 - Default values for optional parameters are no longer being emitted for explicit implementations (issue [#225](https://github.com/JasonBock/Rocks/issues/225))
 - Exact matches for methods on an interface are no longer duplicated (issue [#226](https://github.com/JasonBock/Rocks/issues/226))
 - Handled a case where a type parameter value led to matching members (issue [#217](https://github.com/JasonBock/Rocks/issues/217))